Move LLM translator classes to translations.integrations.llm
Code move only, no behavior change. PoTranslationFormat and the
management command stay put; the provider classes and TranslationFormat
ABC become shared infrastructure for AI app translations.
