+++ corehq/apps/translations/tests/test_translate_po_files.py
@@ -7,11 +7,13 @@ from unittest.mock import MagicMock, patch
 import polib
 import pytest
 
-from corehq.apps.translations.management.commands.translate_po_files import (
+from corehq.apps.translations.integrations.llm import (
     LLMTranslator,
     OpenaiTranslator,
-    PoTranslationFormat,
     TranslationFormat,
+)
+from corehq.apps.translations.management.commands.translate_po_files import (
+    PoTranslationFormat,
 )
 from corehq.tests.tools import nottest
 
@@ -60,7 +62,7 @@ def test_llm_translator_base_prompt():
     )
 
     with patch(
-        'corehq.apps.translations.management.commands.translate_po_files.langcode_to_langname_map'
+        'corehq.apps.translations.integrations.llm.langcode_to_langname_map'
     ) as mock_lang_map:
         mock_lang_map.return_value = {"fra": "French"}
         prompt = translator.base_prompt()
@@ -78,7 +80,7 @@ def test_llm_translator_base_prompt_with_unsupported_lang():
     )
 
     with patch(
-        'corehq.apps.translations.management.commands.translate_po_files.langcode_to_langname_map'
+        'corehq.apps.translations.integrations.llm.langcode_to_langname_map'
     ) as mock_lang_map:
         mock_lang_map.return_value = {"fra": "French"}
         prompt = translator.base_prompt()
@@ -132,7 +134,7 @@ def test_openai_translator_call_llm():
         translator.client.chat.completions.create.assert_called_once()
 
 
-@patch('corehq.apps.translations.management.commands.translate_po_files.requests.post')
+@patch('corehq.apps.translations.integrations.llm.requests.post')
 def test_openai_translator_call_llm_http(mock_post):
     translation_format = MockTranslationFormat()
     translator = OpenaiTranslator(
@@ -154,7 +156,7 @@ def test_openai_translator_call_llm_http(mock_post):
     mock_post.assert_called_once()
 
 
-@patch('corehq.apps.translations.management.commands.translate_po_files.requests.post')
+@patch('corehq.apps.translations.integrations.llm.requests.post')
 def test_openai_translator_client_fallback_to_http(mock_post):
     translation_format = MockTranslationFormat()
     translator = OpenaiTranslator(
